requirements.txt
django==4.0.2
djangorestframework==3.13.1
django-debug-toolbar==3.2.4
